style(View.kt) : inputView에서 View로 클래스 네이밍 수정
